==> notebooks/experiment_2/experiment_2b/step2_featurization_exp2b.ipynb <==
from google.colab import drive
drive.mount('/content/drive')

# define the main train and test dataset path
MAIN_TRAIN_DATA_PATH = "/content/drive/MyDrive/1:1_Shriya_Wagholikar/Datasets/train_data.csv"
MAIN_TEST_DATA_PATH = "/content/drive/MyDrive/1:1_Shriya_Wagholikar/Datasets/test_data.csv"

import pandas as pd
import numpy as np
# read the main train data
train_df = pd.read_csv(MAIN_TRAIN_DATA_PATH)
train_df.head()

# read the main test data
test_df = pd.read_csv(MAIN_TEST_DATA_PATH)
test_df.head()

from sklearn.feature_selection import RFE
from sklearn.ensemble import RandomForestClassifier
# define the components
COMPONENTS = 50
# Estimator
model = RandomForestClassifier()
# select x and y values
X = train_df.drop(columns=['Disease Free (Months)', 'DFS_STATUS_ENCODED'])
y = train_df['DFS_STATUS_ENCODED']
# Apply t-SNE
selector = RFE(model, n_features_to_select=COMPONENTS)
# Select the top 50 features
X.columns = X.columns.astype(str)
X_new = selector.fit_transform(X, y)
# filter the dataframe with the selected features
selected_features = list(X.columns[selector.get_support()]) + ['Disease Free (Months)', 'DFS_STATUS_ENCODED']
print(selected_features)
# select training and testing data based on the filter
train_df_final = train_df[selected_features]
test_df_final = test_df[selected_features]

# define the train and test path
TRAIN_PATH = "/content/drive/MyDrive/1:1_Shriya_Wagholikar/Datasets/experiment_2b/exp2b_rfe_train.csv"
TEST_PATH = "/content/drive/MyDrive/1:1_Shriya_Wagholikar/Datasets/experiment_2b/exp2b_rfe_test.csv"
# save the train csv
train_df_final.reset_index(drop = True, inplace = True)
train_df_final.to_csv(TRAIN_PATH, index = False)
# save the test csv
test_df_final.reset_index(drop = True, inplace = True)
test_df_final.to_csv(TEST_PATH, index = False)
