==> notebooks/experiment_2/experiment_2b/step3_modeltraining_exp2b.ipynb <==
from google.colab import drive
drive.mount('/content/drive')

!pip install lifelines -q

TRAIN_PATH = "/content/drive/MyDrive/1:1_Shriya_Wagholikar/Datasets/experiment_2b/exp2b_rfe_train.csv"
TEST_PATH = "/content/drive/MyDrive/1:1_Shriya_Wagholikar/Datasets/experiment_2b/exp2b_rfe_test.csv"

import pandas as pd
# load the training data
train_data = pd.read_csv(TRAIN_PATH)
train_data.head()

train_data.isnull().sum()

# read the test data
import pandas as pd
test_data = pd.read_csv(TEST_PATH)
test_data.head()

# # we need to drop some columns due to low variance, otherwise convergence failes
# # this was identified by running the model and if there is a converegence error
# # it will show you what columns have lower variance and you can drop them

# COLS_TO_DROP = ['FETUB', 'KRT38'] #provide the column name suggested by model
# #drop those columns from both train and test data
# train_data.drop(COLS_TO_DROP, axis = 1, inplace = True)
# test_data.drop(COLS_TO_DROP, axis = 1, inplace = True)

from sklearn.model_selection import train_test_split, KFold
from lifelines import CoxPHFitter
from lifelines.utils import concordance_index
import numpy as np
def apply_cross_validation(df:pd.DataFrame, penalty: float = 0.1) -> None:
    # create the folds
    kf = KFold(n_splits=5, shuffle=True, random_state=42)
    c_index_scores = []
    # run cross validation
    for train_index, val_index in kf.split(train_data):
        train_fold = df.iloc[train_index]
        val_fold = df.iloc[val_index]
        cph = CoxPHFitter(penalizer = penalty)
        cph.fit(train_fold, duration_col='Disease Free (Months)', event_col='DFS_STATUS_ENCODED')
        val_predictions = cph.predict_partial_hazard(val_fold)
        c_index = concordance_index(val_fold['Disease Free (Months)'], -val_predictions, val_fold['DFS_STATUS_ENCODED'])
        c_index_scores.append(c_index)
    print(f"Average Validation C-index: {np.mean(c_index_scores)}")

apply_cross_validation(train_data, penalty = 0) #find best penalizer value

# Train the final model on the entire training set
cph_final = CoxPHFitter(penalizer = ) #add penalizer after figuring out best value
# fit on train data
cph_final.fit(train_data, duration_col='Disease Free (Months)', event_col='DFS_STATUS_ENCODED')
cph_final.print_summary()

# Concordance Index
c_index = cph_final.concordance_index_
print(f"Concordance Index: {c_index}")

import pickle
MODEL_PATH = "/content/drive/MyDrive/1:1_Shriya_Wagholikar/Results/experiment_2b/FE_RFE_survival_model"
with open(MODEL_PATH, "wb") as pickle_file:
    pickle.dump(cph_final, pickle_file)

# define the train and test path
UPDATED_TRAIN_PATH = "/content/drive/MyDrive/1:1_Shriya_Wagholikar/Datasets/experiment_2b/updated_exp2b_rfe_train.csv"
UPDATED_TEST_PATH = "/content/drive/MyDrive/1:1_Shriya_Wagholikar/Datasets/experiment_2b/updated_exp2b_rfe_test.csv"
train_data.to_csv(UPDATED_TRAIN_PATH, index=False)
test_data.to_csv(UPDATED_TEST_PATH, index=False)
